fix: wrapper class error

+++ dart_native/lib/src/android/foundation/wrapperclass/short.dart
@@ -17,7 +17,7 @@ class JShort extends JSubclass<int> {
 /// New native 'Short'.
 Pointer<Void> _new(dynamic value, String clsName) {
   if (value is int) {
-    JObject object = JObject.parameterConstructor(clsName, [value]);
+    JObject object = JObject.parameterConstructor(clsName, [short(value)]);
     return object.pointer.cast<Void>();
   } else {
     throw 'Invalid param when initializing Short.';
